add logger
