+++ src/app.ts
@@ -1,8 +1,8 @@
 import { setupServer, connectDB } from './server';
 import basicAuth from 'express-basic-auth';
+import logger from './helpers/logging';
+import loggingMiddleware from './middleware/logging.middleware';
 
-import express from 'express';
-import responseTime from 'response-time'
 
 const app = setupServer(false);
 
@@ -19,20 +19,7 @@ app.use('/logs*', basicAuth({
   unauthorizedResponse: getUnauthorizedResponse
 }))
 
-const {createLogger, format, transports} = require('winston');
-
-const logger = createLogger({
-  format: format.combine(
-      format.timestamp(),
-      format.json()
-  ),
-  transports: [
-      new (transports.File)({
-          filename: 'winston.log'
-      }),
-      new (transports.Console)()
-  ]
-});
+app.use(loggingMiddleware);
 
 require('winston-visualize')(app, logger);
 
@@ -40,34 +27,11 @@ process.on("uncaughtExceptionMonitor", (err)=>{
   logger.log('error', err.toString());
 });
 
+
 /**
  * End important logging code - please look above for instructions
  */
 
-app.use(
-  responseTime((req: express.Request, res: express.Response, time: number) => {
-    if (res.statusCode > 350) {
-      logger.log('info', `${req.method} ${req.originalUrl}`, {
-        path: req.originalUrl,
-        statusCode: <number>res.statusCode,
-        time: time,
-        method: req.method,
-        payload: {
-          body: req.body,
-          query: req.query
-        }
-      });
-      return;
-    }
-
-    logger.log('info', `${req.method} ${req.originalUrl}`, {
-      path: req.originalUrl,
-      statusCode: <number>res.statusCode,
-      time: time,
-      method: req.method
-    });
-  })
-);
 
 connectDB(false, logger);
