+++ src/middleware/logging.middleware.ts
@@ -1,0 +1,26 @@
+import logger from '../helpers/logging'
+import responseTime from 'response-time'
+import express from 'express';
+
+export default responseTime((req: express.Request, res: express.Response, time: number) => {
+    if (res.statusCode > 350) {
+        logger.log('info', `${req.method} ${req.originalUrl}`, {
+            path: req.originalUrl,
+            statusCode: <number>res.statusCode,
+            time: time,
+            method: req.method,
+            payload: {
+                body: req.body,
+                query: req.query
+            }
+        });
+        return;
+    }
+
+    logger.log('info', `${req.method} ${req.originalUrl}`, {
+        path: req.originalUrl,
+        statusCode: <number>res.statusCode,
+        time: time,
+        method: req.method
+    });
+});
